requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
fake-useragent==1.1.3
idna==3.4
lxml==4.9.2
pyTelegramBotAPI==4.10.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
python-dotenv~=1.0.0